requirements.txt
pypdf>=3.17.0
Pillow>=10.0.0
reportlab>=4.0.0
requests>=2.31.0
